+++ app/src/main/java/com/example/exchangeapp/screens/UIModels.kt
@@ -13,6 +13,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -48,6 +49,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.exchangeapp.model.service.module.ConnectionStatus
+import com.example.exchangeapp.model.service.module.Message
 import com.example.exchangeapp.model.service.module.currentConnectivityStatus
 import com.example.exchangeapp.model.service.module.observeConnectivityAsFlow
 
@@ -280,5 +282,28 @@ fun MessageBox(
                 )
             }
         }
+    }
+}
+
+@Composable
+fun MessageBubble(message: Message, isSentByCurrentUser: Boolean) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 4.dp)
+            .wrapContentWidth(if (isSentByCurrentUser) Alignment.End else Alignment.Start)
+    ) {
+        Text(
+            text = message.message,
+            style = MaterialTheme.typography.bodyLarge.copy(),
+            color = if (isSentByCurrentUser) Color.White else Color.Black,
+            textAlign = TextAlign.Start,
+            modifier = Modifier
+                .background(
+                    color = if (isSentByCurrentUser) Color(0xFF4CAF50) else Color(0xFFE0E0E0),
+                    shape = MaterialTheme.shapes.medium
+                )
+                .padding(10.dp)
+        )
     }
 }
