+++ app/src/main/java/com/example/exchangeapp/screens/aichat/AiChatScreen.kt
@@ -1,35 +1,26 @@
 package com.example.exchangeapp.screens.aichat
 
-import android.text.Html
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.isImeVisible
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.CleaningServices
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.exchangeapp.model.service.module.Message
 import com.example.exchangeapp.screens.MessageBox
+import com.example.exchangeapp.screens.MessageBubble
 import com.example.exchangeapp.screens.TopBar
 
 @OptIn(ExperimentalLayoutApi::class)
@@ -104,29 +95,6 @@ fun AiChatScreen(
             isEnabled = isEnabled.value,
             isAiChat = true,
             sendMsgFunAI = { viewModel.sendMessage(it) }
-        )
-    }
-}
-
-@Composable
-fun MessageBubble(message: Message, isSentByCurrentUser: Boolean) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp)
-            .wrapContentWidth(if (isSentByCurrentUser) Alignment.End else Alignment.Start)
-    ) {
-        Text(
-            text = message.message,
-            style = MaterialTheme.typography.bodyLarge.copy(),
-            color = if (isSentByCurrentUser) Color.White else Color.Black,
-            textAlign = TextAlign.Start,
-            modifier = Modifier
-                .background(
-                    color = if (isSentByCurrentUser) Color(0xFF4CAF50) else Color(0xFFE0E0E0),
-                    shape = MaterialTheme.shapes.medium
-                )
-                .padding(10.dp)
         )
     }
 }
