+++ app/src/main/java/com/example/exchangeapp/screens/chat/ChatScreen.kt
@@ -4,26 +4,22 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.selection.TextSelectionColors
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.Send
 import androidx.compose.material3.*
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.example.exchangeapp.model.service.module.Message
 import com.example.exchangeapp.screens.MessageBox
+import com.example.exchangeapp.screens.MessageBubble
 
 @OptIn(ExperimentalLayoutApi::class)
 @Composable
@@ -110,25 +106,3 @@ fun ChatScreen(
     }
 }
 
-@Composable
-fun MessageBubble(message: Message, isSentByCurrentUser: Boolean) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(vertical = 4.dp)
-            .wrapContentWidth(if (isSentByCurrentUser) Alignment.End else Alignment.Start)
-    ) {
-        Text(
-            text = message.message,
-            style = MaterialTheme.typography.bodyLarge.copy(),
-            color = if (isSentByCurrentUser) Color.White else Color.Black,
-            textAlign = TextAlign.Start,
-            modifier = Modifier
-                .background(
-                    color = if (isSentByCurrentUser) Color(0xFF4CAF50) else Color(0xFFE0E0E0),
-                    shape = MaterialTheme.shapes.medium
-                )
-                .padding(10.dp)
-        )
-    }
-}
